requirements.txt
discord.py==1.6.0
boto3==1.16.51
